chore: remove comments, refactor automation
